pyproject.toml
[build-system]
requires      = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "teflm"
dynamic = ["version"]
description = "Unofficial implementation of arxiv 2302.4222"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.36.0",
    "open-clip-torch>=3.2.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[project.scripts]
teflm = "teflm.main:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["teflm*"]
exclude = ["tests*"]

[tool.setuptools_scm]
write_to = "_version.py"

[tool.uv.sources]
nnll = { git = "https://github.com/darkshapes/nnll" }
